Updated the methods for the requests to be more clean and also uses static methods in the new Wraparr.Methods namespace. Just helps keeps the code clean and the Reqest just takes a generic object which helps with keeping the code clean and not having to make a method for each type required.

+++ WrappArr/Classes/General/Statistics.cs
@@ -8,6 +8,6 @@ namespace WrappArr.Classes.General
         public int TotalEpisodeCount { get; set; }
         public long SizeOnDisk { get; set; }
         public List<string> ReleaseGroups { get; set; }
-        public int PercentOfEpisodes { get; set; }
+        public double PercentOfEpisodes { get; set; }
     }
 }
